Organize work mapper into work path for clarity when adding collection
