+++ app/jobs/deposit_work_job.rb
@@ -1,13 +1,13 @@
 # frozen_string_literal: true
 
-# Performs a deposit (without SDR API).
-class DepositJob < ApplicationJob
+# Performs a deposit or a work (without SDR API).
+class DepositWorkJob < ApplicationJob
   # @param [WorkForm] work_form
   # @param [Work] work
   # @param [Boolean] deposit if true, deposit the work; otherwise, leave as draft
   def perform(work_form:, work:, deposit:)
     content = Content.find(work_form.content_id)
-    cocina_object = ToCocina::Mapper.call(work_form:, content:, source_id: "h3:object-#{work.id}")
+    cocina_object = ToCocina::Work::Mapper.call(work_form:, content:, source_id: "h3:object-#{work.id}")
     new_cocina_object = perform_persist(cocina_object:, update: work_form.persisted?)
     druid = new_cocina_object.externalIdentifier
     Sdr::Repository.accession(druid:) if deposit
